+++ src/Settings/ITeamscaleWidgetSettings.d.ts
@@ -5,13 +5,10 @@
 export interface ITeamscaleWidgetSettings {
     teamscaleProject: string;
     tgaTeamscaleProject: string;
-    tsaTeamscaleProject: string;
     useSeparateTgaServer: boolean;
-    useSeparateTsaServer: boolean;
     activeTimeChooser: string;
     startFixedDate: number;
     baselineDays: number;
     tsBaseline: string;
     showTestGapBadge: boolean;
-    showTestSmellBadge: boolean;
 }